+++ src/preflate_container.rs
@@ -19,6 +19,7 @@ use crate::{
     scoped_read::ScopedRead,
     statistical_codec::PredictionEncoder,
     token_predictor::TokenPredictor,
+    zstd_compression::{ZstdCompressContext, ZstdDecompressContext},
 };
 
 const COMPRESSED_WRAPPER_VERSION_1: u8 = 1;
@@ -493,50 +494,6 @@ fn verify_file(filename: &str) {
     assert!(v == recompressed);
 }
 
-/// expands the Zlib compressed streams in the data and then recompresses the result
-/// with Zstd with the maximum level.
-pub fn compress_zstd(
-    zlib_compressed_data: &[u8],
-    loglevel: u32,
-    compression_stats: &mut CompressionStats,
-) -> Result<Vec<u8>> {
-    let mut ctx = ZstdCompressContext::new(PreflateCompressionContext::new(loglevel), 9, false);
-
-    let r = ctx.process_vec(zlib_compressed_data)?;
-
-    *compression_stats = ctx.stats();
-
-    Ok(r)
-}
-
-/// decompresses the Zstd compressed data and then recompresses the result back
-/// to the original Zlib compressed streams.
-pub fn decompress_zstd(compressed_data: &[u8], capacity: usize) -> Result<Vec<u8>> {
-    let mut ctx = ZstdDecompressContext::<RecreateFromChunksContext>::new(
-        RecreateFromChunksContext::new(capacity),
-    );
-
-    Ok(ctx.process_vec(compressed_data)?)
-}
-
-#[test]
-fn verify_zip_compress_zstd() {
-    use crate::process::read_file;
-    let v = read_file("samplezip.zip");
-
-    let mut stats = CompressionStats::default();
-    let compressed = compress_zstd(&v, 1, &mut stats).unwrap();
-
-    let recreated = decompress_zstd(&compressed, 256 * 1024 * 1024).unwrap();
-
-    assert!(v == recreated);
-    println!(
-        "original zip = {} bytes, recompressed zip = {} bytes",
-        v.len(),
-        compressed.len()
-    );
-}
-
 #[test]
 fn verify_roundtrip_assert() {
     use crate::process::read_file;
@@ -647,22 +604,6 @@ pub struct PreflateCompressionContext {
     input_complete: bool,
 }
 
-/// used to measure the length of the output without storing it anyway
-struct MeasureWriteSink {
-    pub length: usize,
-}
-
-impl Write for MeasureWriteSink {
-    fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
-        self.length += buf.len();
-        Ok(buf.len())
-    }
-
-    fn flush(&mut self) -> std::io::Result<()> {
-        Ok(())
-    }
-}
-
 impl PreflateCompressionContext {
     pub fn new(log_level: u32) -> Self {
         PreflateCompressionContext {
@@ -733,7 +674,7 @@ impl ProcessBuffer for PreflateCompressionContext {
 }
 
 /// writes the pending output to the writer
-fn write_dequeue(
+pub fn write_dequeue(
     pending_output: &mut VecDeque<u8>,
     writer: &mut impl Write,
     max_output_write: usize,
@@ -778,170 +719,6 @@ impl ProcessBuffer for NopProcessBuffer {
         write_dequeue(&mut self.result, writer, max_output_write).context()?;
 
         Ok(input_complete && self.result.len() == 0)
-    }
-}
-
-pub struct ZstdCompressContext<D: ProcessBuffer> {
-    zstd_compress: zstd::stream::write::Encoder<'static, VecDeque<u8>>,
-    input_complete: bool,
-    internal: D,
-
-    /// if set, the encoder will write all the input to a null zstd encoder to see how much
-    /// compression we would get if we just used Zstandard without any Preflate processing.
-    ///
-    /// This gives a fairer comparison of the compression ratio of Preflate + Zstandard vs. Zstandard
-    /// since Zstd does compress the data a bit, especially if there is a lot of non-Deflate streams
-    /// in the file.
-    test_baseline: Option<zstd::stream::write::Encoder<'static, MeasureWriteSink>>,
-
-    zstd_baseline_size: u64,
-    uncompressed_size: u64,
-    zstd_compressed_size: u64,
-
-    done_write: bool,
-}
-
-impl<D: ProcessBuffer> ZstdCompressContext<D> {
-    pub fn new(internal: D, compression_level: i32, test_baseline: bool) -> Self {
-        ZstdCompressContext {
-            zstd_compress: zstd::stream::write::Encoder::new(VecDeque::new(), compression_level)
-                .unwrap(),
-            input_complete: false,
-            done_write: false,
-            internal,
-            zstd_baseline_size: 0,
-            uncompressed_size: 0,
-            zstd_compressed_size: 0,
-            test_baseline: if test_baseline {
-                Some(
-                    zstd::stream::write::Encoder::new(
-                        MeasureWriteSink { length: 0 },
-                        compression_level,
-                    )
-                    .unwrap(),
-                )
-            } else {
-                None
-            },
-        }
-    }
-}
-
-impl<D: ProcessBuffer> ProcessBuffer for ZstdCompressContext<D> {
-    fn process_buffer(
-        &mut self,
-        input: &[u8],
-        input_complete: bool,
-        writer: &mut impl Write,
-        max_output_write: usize,
-    ) -> Result<bool> {
-        if self.input_complete && (input.len() > 0 || !input_complete) {
-            return Err(PreflateError::new(
-                ExitCode::InvalidParameter,
-                "more data provided after input_complete signaled",
-            ));
-        }
-
-        if input.len() > 0 {
-            self.uncompressed_size += input.len() as u64;
-
-            if let Some(encoder) = &mut self.test_baseline {
-                encoder.write_all(input).context()?;
-            }
-        }
-
-        let done_write = self
-            .internal
-            .process_buffer(
-                input,
-                input_complete,
-                &mut self.zstd_compress,
-                max_output_write,
-            )
-            .context()?;
-
-        if done_write && !self.done_write {
-            self.done_write = true;
-            self.zstd_compress.flush().context()?;
-
-            if let Some(encoder) = &mut self.test_baseline {
-                encoder.do_finish()?;
-                self.zstd_baseline_size = encoder.get_mut().length as u64;
-            }
-        }
-
-        let output = self.zstd_compress.get_mut();
-        let amount_written = write_dequeue(output, writer, max_output_write).context()?;
-        self.zstd_compressed_size += amount_written as u64;
-
-        Ok(done_write && output.len() == 0)
-    }
-
-    fn stats(&self) -> CompressionStats {
-        let s = self.internal.stats();
-        CompressionStats {
-            zstd_compressed_size: self.zstd_compressed_size,
-            zstd_baseline_size: self.zstd_baseline_size,
-            uncompressed_size: self.uncompressed_size,
-            ..self.internal.stats()
-        }
-    }
-}
-
-pub struct ZstdDecompressContext<D: ProcessBuffer> {
-    zstd_decompress: zstd::stream::write::Decoder<'static, VecDeque<u8>>,
-    input_complete: bool,
-    internal: D,
-}
-
-impl<D: ProcessBuffer> ZstdDecompressContext<D> {
-    pub fn new(internal: D) -> Self {
-        ZstdDecompressContext {
-            zstd_decompress: zstd::stream::write::Decoder::new(VecDeque::new()).unwrap(),
-            input_complete: false,
-            internal,
-        }
-    }
-}
-
-impl<D: ProcessBuffer> ProcessBuffer for ZstdDecompressContext<D> {
-    fn process_buffer(
-        &mut self,
-        input: &[u8],
-        input_complete: bool,
-        writer: &mut impl Write,
-        max_output_write: usize,
-    ) -> Result<bool> {
-        if self.input_complete && (input.len() > 0 || !input_complete) {
-            return Err(PreflateError::new(
-                ExitCode::InvalidParameter,
-                "more data provided after input_complete signaled",
-            ));
-        }
-
-        if input.len() > 0 {
-            self.zstd_decompress.write_all(input).context()?;
-        }
-
-        if input_complete && !self.input_complete {
-            self.input_complete = true;
-            self.zstd_decompress.flush().context()?;
-        }
-
-        let output = self.zstd_decompress.get_mut();
-        let slice0 = output.as_slices().0;
-        let is_complete = input_complete && output.len() == slice0.len();
-
-        let r = self.internal.process_buffer(
-            output.as_slices().0,
-            is_complete,
-            writer,
-            max_output_write,
-        );
-
-        output.drain(..slice0.len());
-
-        r
     }
 }
